Move dep5 logic into global_licensing
As detailed in a TODO comment, this makes it easier to share some common
logic between .reuse/dep5 and REUSE.toml.
