+++ src/reuse/report.py
@@ -24,13 +24,8 @@ from uuid import uuid4
 from debian.copyright import Copyright
 
 from . import __REUSE_version__, __version__
-from ._util import (
-    _LICENSEREF_PATTERN,
-    _LICENSING,
-    StrPath,
-    _checksum,
-    _parse_dep5,
-)
+from ._util import _LICENSEREF_PATTERN, _LICENSING, StrPath, _checksum
+from .global_licensing import _parse_dep5
 from .project import Project, ReuseInfo
 
 _LOGGER = logging.getLogger(__name__)
